+++ qoqo_qiskit/src/qoqo_qiskit/backend/backend.py
@@ -17,6 +17,7 @@ from typing import Any, Dict, List, Optional, Tuple, Union, cast
 from qiskit import QuantumCircuit
 from qiskit.providers import Backend
 from qiskit.providers.job import Job
+from qiskit_ibm_runtime import Sampler
 from qiskit_aer import AerSimulator
 from qoqo import Circuit, QuantumProgram
 from qoqo.measurements import ClassicalRegister  # type:ignore
@@ -44,6 +45,7 @@ class QoqoQiskitBackend:
             memory (bool): Whether the output will return the actual single shots instead
                            of an equivalent sequence taken from a result summary.
             compilation (bool): Whether the qiskit `compiler` should be used instead of `run`.
+                (DEPRECATED)
 
         Raises:
             TypeError: the input is not a valid Qiskit Backend instance.
@@ -72,7 +74,7 @@ class QoqoQiskitBackend:
 
         (shots, sim_type) = self._handle_simulation_options(run_options, compiled_circuit)
 
-        job = self._job_execution(compiled_circuit, shots)
+        job = self._job_execution([compiled_circuit], shots, sim_type)
 
         return (job, sim_type, output_registers, input_bit_circuit)
 
@@ -118,9 +120,9 @@ class QoqoQiskitBackend:
             input_bit_circuits_list.append(input_bit_circuit)
             output_registers_list.append(output_registers)
 
-        job = self._job_execution(compiled_circuits_list, cast(int, shots_list))
+        job = self._job_execution(compiled_circuits_list, cast("int", shots_list), sim_type_list)
 
-        return (job, cast(str, sim_type_list), output_registers_list, input_bit_circuits_list)
+        return (job, cast("str", sim_type_list), output_registers_list, input_bit_circuits_list)
 
     def _set_up_registers(
         self,
@@ -136,13 +138,13 @@ class QoqoQiskitBackend:
             output_registers.float_regs_lengths[float_def.name()] = float_def.length()
             if float_def.is_output():
                 output_registers.registers.float_register_dict[float_def.name()] = cast(
-                    List[List[float]], []
+                    "List[List[float]]", []
                 )
         for complex_def in circuit.filter_by_tag("DefinitionComplex"):
             output_registers.complex_regs_lengths[complex_def.name()] = complex_def.length()
             if complex_def.is_output():
                 output_registers.registers.complex_register_dict[complex_def.name()] = cast(
-                    List[List[complex]], []
+                    "List[List[complex]]", []
                 )
         return output_registers
 
@@ -220,10 +222,11 @@ class QoqoQiskitBackend:
         custom_shots = 0
         sim_type = "automatic"
         if run_options["SimulationInfo"]["PragmaGetStateVector"]:
-            compiled_circuit.save_statevector()
+            # compiled_circuit.save_statevector() # noqa: ERA001
+            compiled_circuit.measure_all()
             sim_type = "statevector"
         elif run_options["SimulationInfo"]["PragmaGetDensityMatrix"]:
-            compiled_circuit.save_density_matrix()
+            # compiled_circuit.save_density_matrix() # noqa: ERA001
             sim_type = "density_matrix"
         if "PragmaRepeatedMeasurement" in run_options["MeasurementInfo"]:
             for el in run_options["MeasurementInfo"]["PragmaRepeatedMeasurement"]:
@@ -238,11 +241,12 @@ class QoqoQiskitBackend:
         return shots, sim_type
 
     def _job_execution(
-        self,
-        input_to_send: Union[Circuit, List[Circuit]],
-        shots: int,
+        self, input_to_send: List[Circuit], shots: int, _sim_type: Optional[str]
     ) -> Job:
-        job = self.qiskit_backend.run(input_to_send, shots=shots, memory=self.memory)
+        # job =
+        # self.qiskit_backend.run(input_to_send, shots=shots, memory=self.memory) # noqa: ERA001
+        sampler = Sampler(self.qiskit_backend)
+        job = sampler.run(input_to_send, shots=shots)
         return job
 
     def run_circuit(
@@ -576,7 +580,7 @@ class QoqoQiskitBackend:
                 )
             queued_runs.append(self.run_measurement_queued(program.measurement()))
         elif isinstance(params_values[0], list):
-            params_values = cast(List[List[float]], params_values)
+            params_values = cast("List[List[float]]", params_values)
             for params in params_values:
                 if len(params) != len(input_parameter_names):
                     raise ValueError(