+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This changelog tracks changes of the qoqo_qiskit project starting at version 0.1.0 (initial release).
 
+## 0.10.0
+
+### Changed in 0.10.0
+
+* Migrated `backend.run()` method to using qiskit's `Sampler`
+* Temporarily removed support for statevector simulation (no support for `PragmaGetStateVector` and `PragmaGetDensityMatrix`)
+
 ## 0.9.0
 
 ### Fixed in 0.9.0